Move `FeedImagePresenter` to production module
